refactor: moved commandline utils into separate file
