+++ qri/client.py
@@ -1,60 +1,17 @@
 #!/usr/bin/python
 """ A python client for interacting with qri datasets"""
-from subprocess import Popen, PIPE
+
 import json
 import os
 import re
-import shlex
-import time
 
 import pandas as pd
 
+from .cmd_util import shell_exec, clean_up_files
 
 _HASH_PATTERN = re.compile(r'^\/\w+\/(\w*)')
 _TMP_PATH = os.environ.get("HOME")
-_MAX_ATTEMPTS = 10
-_DELAY = .1
 
-
-#---------------------------------------------------------------------
-def _shell_exec_once(command, cwd=None):
-  """ helper function to execute bash commands and return output"""
-  if cwd:
-    proc = Popen(shlex.split(command),
-                 stdin=PIPE,
-                 stdout=PIPE,
-                 stderr=PIPE,
-                 cwd=cwd)
-    stdoutdata, err = proc.communicate()
-    return stdoutdata, err
-  # else
-  proc = Popen(shlex.split(command), stdin=PIPE, stdout=PIPE, stderr=PIPE)
-  stdoutdata, err = proc.communicate()
-  return stdoutdata.decode(), err
-
-def _shell_exec(command):
-  """ helper function to deal with unreliable commands that may
-      temporarily fail but succeed on a repeated attempt"""
-  stdoutdata, _ = _shell_exec_once(command)
-  for _ in range(_MAX_ATTEMPTS - 1):
-    if "error" not in stdoutdata[:15]:
-      break
-    time.sleep(_DELAY)
-    stdoutdata, _ = _shell_exec_once(command)
-  return stdoutdata
-
-#---------------------------------------------------------------------
-def clean_up_files(paths):
-  """ deletes the files listed in paths """
-  for path in paths:
-    cmd = "rm -rf {}".format(path)
-    _shell_exec(cmd)
-
-def strip_color(colored_text):
-  """ strips color characters to return plaintext for when the
-      global color flag cannot be overriden"""
-  ansi_escape = re.compile(r'\x1B\[[0-?]*[ -/]*[@-~]')
-  return ansi_escape.sub('', colored_text)
 #---------------------------------------------------------------------
 
 def shift_root_down(parent_dict):
@@ -127,7 +84,7 @@ class QriDataset(object):
     cmd = ("qri save {publish}"
            "--file '{head_path}' "
            "-t '{commit_msg}' {name}").format(**flags)
-    result = _shell_exec(cmd)
+    result = shell_exec(cmd)
     print(result)
     # remove tmp files
     # clean_up_files([tmp_body_save_path, tmp_head_save_path])
@@ -161,7 +118,7 @@ def _load_ds_body(name, fix_csv=False):
   """ helper function, loads the data/ 'body' of the dataset """
   flags = {"name": name}
   cmd = "qri body -a {name}".format(**flags)
-  result = _shell_exec(cmd)
+  result = shell_exec(cmd)
   data = json.loads(result)
   if fix_csv:
     data = convert_csv(data)
@@ -171,7 +128,7 @@ def _load_ds_head(name):
   """ helper function, loads the metadata/ 'head' of the dataset"""
   flags = {"name": name}
   cmd = "qri get -f json {name}".format(**flags)
-  result = _shell_exec(cmd)
+  result = shell_exec(cmd)
   parent_dict = json.loads(result)
   return shift_root_down(parent_dict)
 
@@ -192,7 +149,7 @@ def _list_ds(limit, offset):
   """ helper function for `list_ds()`"""
   flags = {"limit": limit, "offset": offset, "format": "json"}
   cmd = "qri -l {limit} -o {offset} -f {format} list".format(**flags)
-  result = _shell_exec(cmd)
+  result = shell_exec(cmd)
   return json.loads(result)
 
 def list_ds(limit=25, offset=0, name_only=True):
